requirements.txt
fastapi
facebook-sdk
pymongo
uvicorn